+++ importd/__init__.py
@@ -2,6 +2,7 @@
 
 
 # stdlib imports
+import copy
 import inspect
 import os
 import sys
@@ -85,6 +86,12 @@ class SmartReturnMiddleware(object):
 
 
 class Blueprint(object):
+    """
+    Blueprint is way to group urls.
+    This class is used for save blueprint info.
+    The instance of blueprint class is used inside D object initialization.
+    """
+
     def __init__(self):
         self.url_prefix = None
         self.namespace = None
@@ -510,14 +517,19 @@ class D(object):
         management.execute_from_command_line([sys.argv[0]] + list(args))
 
     def register_blueprint(self, bp, url_prefix, namespace, app_name=''):
-        from copy import deepcopy
-        clone_bp = deepcopy(bp)
+        """
+        Interface to register blueprint.
+
+        see django url namespace.
+        https://docs.djangoproject.com/en/1.7/topics/http/urls/#url-namespaces
+        """
+        clone_bp = copy.deepcopy(bp)
         clone_bp.url_prefix = url_prefix
         clone_bp.namespace = namespace
         clone_bp.app_name = app_name
         self.blueprint_list.append(clone_bp)
 
-    def apply_blueprint(self, bp):
+    def _apply_blueprint(self, bp):
         try:
             from django.conf.urls import include
         except ImportError:
@@ -539,7 +551,7 @@ class D(object):
 
     def do(self, *args):
         for bp in self.blueprint_list:
-            self.apply_blueprint(bp)
+            self._apply_blueprint(bp)
 
         if not args:
             args = sys.argv[1:]